+++ src/pages/join/components/index.ts
@@ -1,4 +1,8 @@
 export * from './join-stepper'
 export * from './term-form'
-export * from './details-form'
 export * from './info'
+
+export * from './join-form/email-verify-form'
+export * from './join-form/join-field-info'
+export * from './join-form/password-form'
+export * from './join-form/personal-form'
